Move auth algorithm to a dll.
