+++ CPPCallCS/ClassLibrary1/ClassLibrary1.csproj
@@ -44,7 +44,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Authentication.cs" />
     <Compile Include="Class1.cs" />
     <Compile Include="IMyInterface1.cs" />
     <Compile Include="LoginFrm.cs">
@@ -59,6 +58,12 @@
     <EmbeddedResource Include="LoginFrm.resx">
       <DependentUpon>LoginFrm.cs</DependentUpon>
     </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SXPWDLib\SXPWDLib.csproj">
+      <Project>{b4af73cd-7b17-46e1-8444-1693981f054f}</Project>
+      <Name>SXPWDLib</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
