+++ CPPCallCS/PwdGenerater/Form1.cs
@@ -1,4 +1,4 @@
-﻿using SXPWD;
+﻿using SXPWDLib;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
